Code for pulling Steam ID data
This R code creates a list of 100 potential Steam IDs. It then tries to
pull JSON data from those IDs using the Steam API. If the IDs do exist
and are public then it retrieves the JSON data provided and converts it
into a CSV file which it saves. This code repeats itself a designated
number of times.